pipenv/docker/requirements.txt
fastapi
uvicorn
pandas
redis
mysql-connector-python